# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rental Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyType } from "./propertyType";

export interface Property {
  id: number;
  name: string;
  address: string;
  city: string;
  state: string;
  zipCode?: string;
  type: PropertyType;
  totalUnits: number;
  occupiedUnits: number;
  createdAt: Date;
}